ACCUMULO-44 redesigned RowIterator and updated classes that use it
git-svn-id: https://svn.apache.org/repos/asf/incubator/accumulo/trunk@1185680 13f79535-47bb-0310-9956-ffa450edef68

+++ src/core/src/main/java/org/apache/accumulo/core/client/mapreduce/InputFormatBase.java
@@ -571,7 +571,7 @@ public abstract class InputFormatBase<K,V> extends InputFormat<K,V> {
 	
 	
 	protected abstract static class RecordReaderBase<K,V> extends RecordReader<K,V> {
-		protected int numKeysRead;
+		protected long numKeysRead;
 		protected Iterator<Entry<Key, Value>> scannerIterator;
 		private boolean scannerRegexEnabled = false;
 		protected RangeInputSplit split;
